% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hive colors.R
\name{hive_palette}
\alias{hive_palette}
\title{Hive Palette Generator (for ggplot2 discrete scales)}
\usage{
hive_palette(alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{alpha}{Transparency level.}

\item{begin}{Start point.}

\item{end}{End point.}

\item{direction}{Gradient direction.}
}
\value{
Function(n) that generates n hex colors.
}
\description{
Returns a palette function usable inside ggplot2's discrete_scale().
}
