% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hive colors.R
\name{scale_fill_hive}
\alias{scale_fill_hive}
\title{ggplot2 Fill Scale using Hive Colors}
\usage{
scale_fill_hive(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE
)
}
\arguments{
\item{...}{Additional parameters for scale functions.}

\item{alpha}{Transparency.}

\item{begin}{Start point.}

\item{end}{End point.}

\item{direction}{Gradient direction.}

\item{discrete}{If TRUE, uses discrete scale; if FALSE, uses continuous gradient.}
}
\value{
A ggplot2 scale object.
}
\description{
Apply the Hive palette to the ggplot2 fill aesthetic.
}
